+++ docs/CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 - Add filter by sample to variant objects
 - Add additional parameters for ``calculate-bigwig`` process
 - Add ``samtools-mpileup-single`` process
+- Add ``arriba`` process
 
 Changed
 -------